+++ partiql-planner/src/test/kotlin/org/partiql/planner/HeaderTest.kt
@@ -8,6 +8,6 @@ class HeaderTest {
     @Test
     @Disabled
     fun print() {
-        println(PartiQLHeader.toString())
+        println(PartiQLHeader)
     }
 }
